pyproject.toml
[build-system]
requires = ["flit_core >=3.4,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "aiida-muon"
dynamic = ["version"] # read from aiida_muon/__init__.py
description = "aiida-muon is allows to find candidate muon implantation sites and hyperfine field by DFT supercell relaxations and from further symmetry and kinetics analysis.  "
authors = [{name = "Muon group Parma"},{name = "Laboratory for Materials Simulations PSI"}]
readme = "README.md"
license = {file = "LICENSE.txt"}
classifiers = [
    "Programming Language :: Python",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Free To Use But Restricted",
    "Natural Language :: English",
    "Development Status :: 5 - Production/Stable",
    "Framework :: AiiDA"
]
keywords = ["aiida", "workflows"]
requires-python = ">=3.8"
dependencies = [
    "aiida-core>=2.0,<3",
    "aiida-quantumespresso>=4.2.0",
    "aiida-pseudo",
    "pymatgen",
    "spglib>=1.9",
    "pyyaml>=3.0",
    "mulfc @ git+https://github.com/bonfus/muLFC.git@v0.0.4",
    "muesr @ git+https://github.com/bonfus/muesr@master",
    "aiida-impuritysupercellconv @ git+https://github.com/mikibonacci/aiida-impuritysupercellconv@new_release/cleanup"
]

[project.urls]
Source = "https://github.com/positivemuon/aiida-muon"

[project.optional-dependencies]
docs = [
    "sphinx",
    "docutils",
    "sphinx-copybutton~=0.3.0",
    "sphinx-book-theme~=0.1.0",
    "sphinx-click~=2.7.1",
    "myst"
]
pre-commit = [
    "pre-commit>=2.21.0"
]
tests = [
    "pgtest~=1.3",
    "pytest~=6.0",
    "pytest-regressions~=2.3"
]

[project.entry-points."aiida.workflows"]
"muon.find_muon" = "aiida_muon.workflows.find_muon:FindMuonWorkChain"

[tool.flit.module]
name = "aiida_muon"

[tool.pytest.ini_options]
minversion = "6.0"
testpaths = [
    "tests"
]
filterwarnings = [
    "ignore:Creating AiiDA configuration folder.*:UserWarning",
    "ignore::DeprecationWarning:frozendict:",
    "ignore::DeprecationWarning:pkg_resources:",
    "ignore::DeprecationWarning:sqlalchemy_utils:"
]
